requirements.txt
python-telegram-bot==13.15
google-auth==2.22.0
requests==2.31.0
